+++ CHANGELOG.rst
@@ -4,4 +4,5 @@ Unreleased
 * feat: `Introduction.introduction`
 * feat: `Introduction.ombud_messages`
 * feat: `database.boards`
+* feat: `database.cs_codes`
 * fixed: `database.add_instrument_to_db`